Improved code, created a new module (handle_markdowns) and moved a few functions there
